finished custom hook for cards and host refs
