+++ src/components/RoomView/RoomView.tsx
@@ -1,5 +1,4 @@
-import React, { useEffect,useState } from 'react';
-import axios from 'axios';
+import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { database } from '../../utils/firebase';
 import { useParams } from 'react-router-dom';
@@ -8,7 +7,6 @@ import './roomView.css';
 import Popup from '../Room/Popup';
 import { isEveryRoleFilled } from '../../utils/utilFunctions';
 import SetupGame from './SetupGame';
-import WelcomeBoard from '../Navbar/WelcomeBoard';
 import OperativeBoard from './OperativeBoard';
 import SpyMasterBoard from './SpyMasterBoard';
 import TeamOneBox from '../teamBoxes/TeamOneBox';
@@ -31,13 +29,12 @@ import {
 } from '../../store/gameSlice';
 import { setCurrentClue } from '../../store/clueSlice';
 import { RootState } from '../../store/index.js';
-import { CardObj, WordsWithTeamIdsObj } from '../../utils/interfaces';
-import { setHost } from '../../store/gameSlice';
-import { setIsHost } from '../../store/playerSlice';
-import Loser from'./Loser';
-import Winner from './Winner'
-import words from 'random-words';
+
+import Loser from './Loser';
+import Winner from './Winner';
 import Navbar from '../Navbar/Navbar';
+import OnValueHostRef from './customHooks/OnValueHostRef';
+import OnValueCardsRef from './customHooks/OnValueCardsRef';
 
 interface ClassName {
   className: string;
@@ -52,19 +49,14 @@ const RoomView = (props: ClassName) => {
   const [timedPopup, setTimedPopup] = useState(false);
   // frontend state
   const { playerId, username, isHost } = useSelector((state: RootState) => state.player);
-  const { winner, loser} = useSelector((state: RootState) => state.game);
   const { teamOneOperatives, teamOneSpymaster } = useSelector((state: RootState) => state.teamOne);
   const { teamTwoOperatives, teamTwoSpymaster } = useSelector((state: RootState) => state.teamTwo);
   const { host } = useSelector((state: RootState) => state.game);
   // firebase room  & players reference
   let playersInRoomRef = ref(database, 'rooms/' + roomId + '/players/');
   let gameRef = ref(database, 'rooms/' + roomId + '/game/');
-  let cardsRef = ref(database, `rooms/${roomId}/gameboard`);
   let gameHistoryRef = ref(database, `rooms/${roomId}/game/history`);
-  const teamOneSpymasterRef = ref(database, `rooms/${roomId}/team-1/spymaster/`);
-  const teamOneOperativesRef = ref(database, `rooms/${roomId}/team-1/operatives/`);
-  const teamTwoOperativesRef = ref(database, `rooms/${roomId}/team-2/operatives/`);
-  const teamTwoSpymasterRef = ref(database, `rooms/${roomId}/team-2/spymaster/`);
+
   let hostRef = ref(database, `rooms/${roomId}/host`);
 
   // below will be used once we allow host & everyones here to show button
@@ -152,115 +144,11 @@ const RoomView = (props: ClassName) => {
         }
         dispatch(setGameHistory(history));
       }
-    }); setTimeout(() => {
+    });
+    setTimeout(() => {
       setTimedPopup(true);
     }, 1000);
   }, []);
-
-  useEffect(() => {
-    // Look to see if there are cards already loaded for the room
-    onValue(cardsRef, async (cardSnapshot) => {
-      // for some reason, i'm having trouble accessing the redux teams
-      //  data even though it exists on firebase and redux
-      // tried a few diff ways and this is what i could get to work. bulky :(
-      if (cardSnapshot.exists()) {
-        get(teamOneSpymasterRef).then(async (snapshot) => {
-          if (snapshot.exists()) {
-            let spymaster = snapshot.val();
-
-            if (spymaster.playerId === playerId) {
-              //get set of cards with team ids from backend and set spymaster words
-              let wordsWithTeamIds = {} as WordsWithTeamIdsObj;
-              let spyWords = await axios.get(`/api/card/get25/forRoom/${roomId}`);
-              spyWords.data.forEach(
-                (card: CardObj) =>
-                  (wordsWithTeamIds[card.id] = {
-                    id: card.id,
-                    isVisibleToAll: card.isVisibleToAll,
-                    wordString: card.word.word,
-                    word: card.word,
-                    wordId: card.wordId,
-                    boardId: card.boardId,
-                    teamId: card.teamId,
-                  }),
-              );
-              const values = Object.values(wordsWithTeamIds);
-              dispatch(setWordsInGame(values));
-            }
-          }
-        });
-        get(teamTwoSpymasterRef).then(async (snapshot) => {
-          if (snapshot.exists()) {
-            let spymaster = snapshot.val();
-            if (spymaster.playerId === playerId) {
-              console.log('setting spy board...');
-              //get set of cards with team ids from backend and set spymaster words
-              let wordsWithTeamIds = {} as WordsWithTeamIdsObj;
-              let spyWords = await axios.get(`/api/card/get25/forRoom/${roomId}`);
-              spyWords.data.forEach(
-                (card: CardObj) =>
-                  (wordsWithTeamIds[card.id] = {
-                    id: card.id,
-                    isVisibleToAll: card.isVisibleToAll,
-                    wordString: card.word.word,
-                    word: card.word,
-                    wordId: card.wordId,
-                    boardId: card.boardId,
-                    teamId: card.teamId,
-                  }),
-              );
-              const values = Object.values(wordsWithTeamIds);
-              dispatch(setWordsInGame(values));
-            }
-          }
-        });
-        get(teamOneOperativesRef).then((snapshot) => {
-          if (snapshot.exists()) {
-            let operatives = snapshot.val();
-            let operativesIds = Object.keys(operatives);
-            if (operativesIds.includes(playerId)) {
-              console.log('setting opertive board...');
-              //update our redux to reflect that
-              const cardsFromSnapshot = cardSnapshot.val();
-              const values = Object.values(cardsFromSnapshot);
-              dispatch(setWordsInGame(values));
-            }
-          }
-        });
-        get(teamTwoOperativesRef).then((snapshot) => {
-          if (snapshot.exists()) {
-            let operatives = snapshot.val();
-            let operativesIds = Object.keys(operatives);
-            if (operativesIds.includes(playerId)) {
-              console.log('setting opertive board...');
-              //update our redux to reflect that
-              const cardsFromSnapshot = cardSnapshot.val();
-              const values = Object.values(cardsFromSnapshot);
-              dispatch(setWordsInGame(values));
-            }
-          }
-        });
-      }
-    });
-  }, [playerId]);
-
-  useEffect(() => {
-    onValue(hostRef, (snapshot) => {
-      if (snapshot.exists()) {
-        const host = snapshot.val();
-        if (host.playerId === playerId) {
-          dispatch(setHost(host));
-          dispatch(setIsHost(true));
-        } else {
-          dispatch(setHost(host));
-          dispatch(setIsHost(false));
-        }
-      } else {
-        dispatch(setHost(null));
-        dispatch(setIsHost(false));
-      }
-    });
-  }, [playerId]);
 
   // this function works everywhere else without having to 'get' the gamestatus from firebase
   // it would NOT cooperate or pull accurate game status from redux. :|
@@ -289,16 +177,23 @@ const RoomView = (props: ClassName) => {
   const claimHost = () => {
     update(hostRef, { playerId, username });
     update(child(playersInRoomRef, playerId), { playerId, username, isHost: true });
-  }
+  };
+
+  OnValueHostRef();
+  OnValueCardsRef();
   return (
     <div className={`${props.className} roomViewBG`}>
-          <Navbar />
+      <Navbar />
       <div className="gameStatusClaimHost">
-      <GameStatus />
-      <div className='gameStatus'>
-      {!host && 
-      <p>The host has left, need a <button onClick={claimHost}>New Host</button> to begin game.</p>
-      }</div></div>
+        <GameStatus />
+        <div className="gameStatus">
+          {!host && (
+            <p>
+              The host has left, need a <button onClick={claimHost}>New Host</button> to begin game.
+            </p>
+          )}
+        </div>
+      </div>
       {/* is there isnt at least one person to each role, setup board should be disabled / not visible */}
       {/* is host AND there is at least one person on each team */}
       {isHost && (
@@ -324,7 +219,8 @@ const RoomView = (props: ClassName) => {
         <div className="chatBox"> this will be the chat box</div>
       </div>
       <Clue />
-<Loser /><Winner />
+      <Loser />
+      <Winner />
       {/* COMMENTING OUT THE BELOW CODE UNTIL WE'RE READY TO TEST WTH ALL ROLES FILLED */}
       {/* {isHost && everyonesHere && (
         <Popup
@@ -343,8 +239,7 @@ const RoomView = (props: ClassName) => {
           <SetupGame />
         </Popup>
       )} */}
-      
-  </div>)
-
-    }
+    </div>
+  );
+};
 export default RoomView;
